requirements.txt
langchain~=0.3.25
wikipedia
langchain-community~=0.3.25
langchain-openai~=0.3.24
langchain-anthropic
python-dotenv~=1.1.0
pydantic~=2.11.7
duckduckgo-search
openai~=1.88.0
matplotlib~=3.10.3
langchain-core~=0.3.65
requests~=2.32.4
pillow~=11.2.1